1st lab done.

+++ computer-graphics/1st-lab/image-filter/ImageFilter/Abstractions/IRgbConverter.cs
@@ -8,6 +8,6 @@ namespace ImageFilter.Abstractions
 {
     public interface IRgbConverter
     {
-        public void ConvertToRgb(in double[,,] source, out double[,,] rgbValues);
+        public void ConvertToRgb(in double[,,] source, out int[,,] rgbValues);
     }
 }
